% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{plot.DTSg}
\alias{plot.DTSg}
\alias{plot}
\title{Plot time series data}
\usage{
\method{plot}{DTSg}(
  x,
  from = first(self$values(reference = TRUE)[[".dateTime"]]),
  to = last(self$values(reference = TRUE)[[".dateTime"]]),
  cols = self$cols(class = "numeric"),
  secAxisCols = NULL,
  secAxisLabel = "",
  ...
)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{from}{A \code{\link{POSIXct}} timestamp in the same time zone as the time series
or a character string coercible to one. The data is plotted from this
timestamp on.}

\item{to}{A \code{\link{POSIXct}} timestamp in the same time zone as the time series or
a character string coercible to one. The data is plotted up to this
timestamp.}

\item{cols}{A character vector specifying the columns whose values shall be
plotted.}

\item{secAxisCols}{An optional character vector specifying the columns whose
values shall be plotted on a secondary axis. Must be a subset of \code{cols}.}

\item{secAxisLabel}{A character string specifying the label of the optional
secondary axis.}

\item{...}{Not used (S3 method only).}
}
\value{
Returns a \code{DTSg} object.
}
\description{
Displays an interactive plot of a \code{\link{DTSg}} object. This method requires
\pkg{dygraphs} and \pkg{RColorBrewer} to be installed. Its main purpose is
not to make pretty plots, but rather to offer a possibility to interactively
explore the time series data. The title of the plot and the label of its
primary axis are automatically generated from the object's metadata (fields).
See \code{\link{DTSg}} for further information.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# plot data
if (requireNamespace("dygraphs", quietly = TRUE) &&
    requireNamespace("RColorBrewer", quietly = TRUE)) {
  ## R6 method
  x$plot()

  ## S3 method
  plot(x = x)
}

}
\seealso{
\code{\link{cols}}
}
