% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{print.DTSg}
\alias{print.DTSg}
\alias{print}
\title{Print object}
\usage{
\method{print}{DTSg}(x, ...)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{...}{Not used (S3 method only).}
}
\value{
Returns a \code{DTSg} object.
}
\description{
Prints a \code{\link{DTSg}} object.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# print object
## R6 method
x$print()

## S3 method
print(x = x)

}
