% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{rowapply.DTSg}
\alias{rowapply.DTSg}
\alias{rowapply}
\title{Apply Function(s) Row-wise}
\usage{
\method{rowapply}{DTSg}(
  x,
  fun,
  ...,
  resultCols,
  cols = self$cols(class = "numeric"),
  clone = getOption("DTSgClone")
)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{fun}{A \code{\link{function}} or a named \code{\link{list}} of
functions applied row-wise to all the values of the specified columns, for
instance, \code{\link{mean}} or \code{\link{list}(min = \link{min},
max = \link{max})}. The return value(s) must be of length one.}

\item{\dots}{Further arguments passed on to \code{fun}.}

\item{resultCols}{A character vector either of length one (names of
\code{fun} are appended in the case more than one function is provided) or
the same length as \code{fun}.}

\item{cols}{A character vector specifying the columns to apply \code{fun} to.}

\item{clone}{A logical specifying if the object is modified in place or if a
clone (copy) is made beforehand.}
}
\value{
Returns a \code{\link{DTSg}} object.
}
\description{
Applies one or more arbitrary functions row-wise to selected columns of a
 \code{\link{DTSg}} object.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# check object integrity
## R6 method


## S3 method


}
\seealso{
\code{\link{DTSg}}, \code{\link{function}}, \code{\link{list}},
 \code{\link{cols}}
}
