% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{summary.DTSg}
\alias{summary.DTSg}
\alias{summary}
\title{Time Series Summary}
\usage{
\method{summary}{DTSg}(object, cols = self$cols(), ...)
}
\arguments{
\item{object}{A \code{\link{DTSg}} object (S3 method only).}

\item{cols}{A character vector specifying the columns whose values shall be
summarised.}

\item{\dots}{Further arguments passed on to \code{\link{summary.data.frame}}.}
}
\value{
Returns a \code{\link{table}}.
}
\description{
Calculates summary statistics of selected columns of a \code{\link{DTSg}}
 object.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# calculate summary statistics
## R6 method
x$summary()

## S3 method
summary(object = x)

}
\seealso{
\code{\link{DTSg}}, \code{\link{cols}},
 \code{\link{summary.data.frame}}, \code{\link{table}}
}
