% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{cols.DTSg}
\alias{cols.DTSg}
\alias{cols}
\alias{names}
\title{Get column names}
\usage{
\method{cols}{DTSg}(x, class = NULL, pattern = NULL, mode = NULL, typeof = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{class}{An optional character vector matched to the most specific class
(first element) of each column's \code{\link{class}} vector. The \dQuote{special
class} \code{".numerary"} matches the \code{\link{integer}} and \code{\link{numeric}} classes.}

\item{pattern}{An optional character string passed on to the \code{pattern}
argument of \code{\link{grep}}.}

\item{mode}{An optional character vector matched to each column's \code{\link{mode}}.}

\item{typeof}{An optional character vector matched to each column's
\code{\link{typeof}}.}

\item{\dots}{Further arguments passed on to \code{\link{grep}}. The \code{value} argument is
rejected.}
}
\value{
Returns a character vector.
}
\description{
Returns all column names of a \code{\link{DTSg}} object, those of certain \code{\link{class}}es,
\code{\link{mode}}s, \code{\link{typeof}}s and/or those matching a certain pattern only.
}
\section{R6 alias}{

The \code{names()} alias for this method is exclusively available in the R6
interface.
}

\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# get names of numeric columns
## R6 method
x$cols(class = "numeric")

## 'names()' is an R6 alias for 'cols()'
x$names(class = "numeric")

## S3 method
cols(x = x, class = "numeric")

}
