% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{getCol.DTSg}
\alias{getCol.DTSg}
\alias{getCol}
\alias{[.DTSg}
\title{Get Column Vector}
\usage{
\method{getCol}{DTSg}(x, col = self$cols(class = "numeric")[1L], ...)

\method{[}{DTSg}(x, ...)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{col}{A character string specifying a column name.}

\item{\dots}{Passes on arguments from \code{[} to \code{getCol}.
\code{getCol} itself does not make use of it.}
}
\value{
Returns a vector or a \code{\link{list}} in case of a
 \code{\link{list}} column.
}
\description{
Queries the values of a column of a \code{\link{DTSg}} object.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# get values of "flow" column
## R6 methods
x$getCol(col = "flow")
x$`[`("flow")

## S3 methods
getCol(x = x, col = "flow")
x["flow"]

}
\seealso{
\code{\link{DTSg}}, \code{\link{cols}}, \code{\link{list}}
}
