% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{setCols.DTSg}
\alias{setCols.DTSg}
\alias{setCols}
\title{Set Columns' Values}
\usage{
\method{setCols}{DTSg}(x, i, cols = self$cols(class = "numeric")[1L],
  values, clone = getOption("DTSgClone"), ...)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{i}{A numeric vector indexing rows or a filter expression as accepted by
the \code{i} argument of \code{\link[data.table]{data.table}}. Filter
expressions can contain the special symbol \code{.N}. See
(\code{\link[data.table]{special-symbols}}) for further information.}

\item{cols}{A character vector specifying the columns whose values shall be
set. The values of the \emph{.dateTime} column cannot be set.}

\item{values}{A list of replacement and/or new values as accepted by the
\code{value} argument of \code{\link[data.table]{set}}. \code{\link{NULL}}
as value removes a column.}

\item{clone}{A logical specifying if the object is modified in place or if a
clone (copy) is made beforehand.}

\item{\dots}{Not used (S3 method only).}
}
\value{
Returns a \code{\link{DTSg}} object.
}
\description{
Set the values of columns, add columns to and/or remove columns from a
 \code{\link{DTSg}} object. The values can optionally be set for certain rows
 only.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# limit river flow to 100
## R6 method
x$setCols(i = flow > 100, cols = "flow", values = 100)

## S3 method
setCols(x = x, i = flow > 100, cols = "flow", values = 100)

}
\seealso{
\code{\link{DTSg}}, \code{\link[data.table]{data.table}},
 \code{\link[data.table]{special-symbols}}, \code{\link{cols}},
 \code{\link[data.table]{set}}, \code{\link{NULL}}
}
