% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swrappers.R
\name{subset.DTSg}
\alias{subset.DTSg}
\alias{subset}
\title{Subset Time Series}
\usage{
\method{subset}{DTSg}(x, i, cols = self$cols(),
  na.status = self$na.status, clone = getOption("DTSgClone"), ...)
}
\arguments{
\item{x}{A \code{\link{DTSg}} object (S3 method only).}

\item{i}{A numeric or logical vector or an expression as accepted by the
\code{i} argument of \code{\link[data.table]{data.table}}.}

\item{cols}{A character vector specifying the columns to select. The
\emph{.dateTime} column is always selected and cannot be part of it.}

\item{na.status}{A character string. Either \code{"explicit"}, which makes
missing timestamps according to the recognised periodicity explicit, or
\code{"implicit"}, which removes timestamps with missing values on all value
columns. Please note that filtering rows and having or making missing
timestamps explicit equals to setting the values of all the other timestamps
to missing.}

\item{clone}{A logical specifying if the object is modified in place or if a
clone (copy) is made beforehand.}

\item{\dots}{Not used (S3 method only).}
}
\value{
Returns a \code{\link{DTSg}} object.
}
\description{
Filter rows and/or select columns of a \code{\link{DTSg}} object.
}
\examples{
# new DTSg object
x <- DTSg$new(values = flow)

# filter for non-missing river flow between the 280th and 320th row
## R6 method
x$subset(i = 280:320, na.status = "implicit")

## S3 method
subset(x = x, i = 280:320, na.status = "implicit")

}
\seealso{
\code{\link{DTSg}}, \code{\link[data.table]{data.table}},
 \code{\link{cols}}
}
